Update template for parallel use

+++ asr/tests/template.py
@@ -2,6 +2,7 @@
 import asr
 from asr.utils.recipe import Recipe
 from pathlib import Path
+from gpaw.mpi import world
 
 material = '# Material #'
 formula = material.split('.')[0]
@@ -12,10 +13,10 @@ srcparams = Path(asrtestfolder) / f'{formula}_params.json'
 dststart = Path('.') / 'unrelaxed.json'
 dstparams = Path('.') / 'params.json'
 
-if not dststart.is_file():
+if not dststart.is_file() and world.rank == 0:
     dststart.write_bytes(srcstart.read_bytes())
 
-if not dstparams.is_file():
+if not dstparams.is_file() and world.rank == 0:
     dstparams.write_bytes(srcparams.read_bytes())
 
 # Make sure to reload the module
